Move functions to relevant file names.
